+++ crates/cubed-application/src/ports/mod.rs
@@ -1,3 +1,5 @@
+pub mod file_system;
 pub mod server_repository;
 
+pub use file_system::FileSystemManager;
 pub use server_repository::ServerRepository;