+++ crates/cubed-application/src/use_cases/mod.rs
@@ -1,11 +1,13 @@
 pub mod create_server;
 pub mod delete_server;
+pub mod init_file_system;
 pub mod restart_server;
 pub mod start_server;
 pub mod stop_server;
 
 pub use create_server::{CreateServer, CreateServerInput};
 pub use delete_server::DeleteServer;
+pub use init_file_system::InitFileSystem;
 pub use restart_server::RestartServer;
 pub use start_server::StartServer;
 pub use stop_server::StopServer;